Successfully created an entity from a script

+++ src/net/lodoma/lime/client/io/entity/CPHEntityCreation.java
@@ -27,7 +27,7 @@ public class CPHEntityCreation extends ClientPacketHandler
         EntityLoader entityLoader = (EntityLoader) client.getProperty("entityLoader");
         
         int id = inputStream.readInt();
-        long hash = inputStream.readLong();
+        int hash = inputStream.readInt();
         
         try
         {
